Fixed handling of bool values in initrd code
Some values evaluated in the initrd code are created in
the kiwi builder and passed in as a profile file. bool
values created by kiwi for use in shell scripts takes
the string 'true' or 'false' or are not set at all if not
specified in the kiwi XML description. Some code paths
in the initrd code uses the '-n' switch to check for bool
values, however if the string 'false' is passed '-n' will
do the wrong thing. Therefore a method for the initrd
code has been introduced to clearly handle bool values