Allow dracut initrd to be used with pxe type
The pxe image type builds a simple filesystem image plus a
custom initrd. That initrd is usually build from the kiwi
netboot image descriptions which implements a workflow to
fetch the filesystem image and deploy it as rootfs for a
pxe client. User who wants to implement their own handling
of the rootfs image in e.g a custom dracut module were not
able to build this image type because we did not allow
the initrd_system attribute for the pxe type