Make kiwi-overlay module compatible with kiwi-repart
This commit makes the modules provided by dracut-kiwi-overlay and
dracut-kiwi-oem-repart compatible. The main idea of this commit
is forcing the preparation of the overlay rootfs write partition
after kiwi-repart gets into the play. Since kiwi-overlay mounts the
write partitiontition it has to be executed after the kiwi-repart.

For that instead of doing the job using an initqueue event added by a
udev rule we just add the very same logic in pre-mount hook that happens
after the kiwi-repart pre-mount hook thanks to lower priority.

Fixes #1184