Support specifying the source type of repos
Repository source paths can be more than just a baseurl.
There is also the opportunity to point to a metalink
file or a mirrorlist. In both cases the packagemanager
needs to know the type of this target. Therefore a new
attribute called sourcetype exists which allows to
specify the type the repo path uri is associated with.
Currently the dnf repository class is the only one that
makes use of the information. This change is required to
support Fedora >= 30 images. I also expect this concept
to be used by other distributions in the future