Delete use of linuxefi/initrdefi grub commands
The above commands were used by kiwi on x86 in EFI mode but
seems to be deleted from upstream grub and therefore tend
to be no longer present on the distributions. In addition
my tests with older distributions and simple use of the
linux/initrd standard commands in EFI mode were successful
which lead to this patch.