Complete dracut setup for luks boot
An image that is configured with an encrypted root including /boot
includes a /.root.keyfile in initrd to let dracut/systemd decrypt
the root and mount it without asking the password. On rebuild of
the initrd, dracut has no configuration that tells it to include
the /.root.keyfile again. This patch adds that configuration and
Fixes #1192