Fixed rpmdb compat link setup
On older versions of zypper the path /var/lib/rpm was hardcoded
and not used from the rpm macro definition. For such systems and
to support them properly on hosts that have the rpm database
already moved a compat link was created. However if the host has
the rpm database at /var/lib/rpm the link doesn't make sense.
This patch fixes this and therefore bsc#1150190