Adapted handling of initrd boot names
Due to the change of using grub2-mkconfig the way custom kiwi
initrd file names are handled doesn't work anymore. grub2-mkconfig
is not able to cope with the names kiwi used when creating custom
kiwi initrd's (initrd.vmx). Thus the same naming code as used
for the dracut initrd system has to apply for the custom kiwi
initrd code as well. This commit moves the name handling into
the baseclass because the same naming schema now applies to
both initrd systems.