Move suse tests to obsrepositories
Changes on the obs backend made this change necessary.
The chain of process decisions is different if a repo is
added to the prj config compared to if the repo is added
to the kiwi XML file. It seems the preferred solution
is the obs prj config and I'm tired running into problems
again and again