Adapt test-image-oem integration test for s390
The test was originally designed to test for DASD 4k block
storage. However the kpartx utility in the Leap15, TW code
stream has issues mapping partitions if the loop device
was setup using 4k sector size. So far we can't create
images with 4k blocksize due to that issue. Thus the
integration test is now adapted for an emulated DASD device
in FBA mode which is not using 4k blocksize. Once the
problem with kpartx is solved on s390 we will create another
integration test to test 4k image builds