Fixed subsequent reboot of oem disk
On a second reboot of an oem disk we check with gdisk's
verification command if the disk needs to be resized.
That command however mounts the disk in the background
and therefore it's urgently required to mask the systemd
rootfs service before. Otherwise systemd thinks this is
evil and drops into a rescue shell