Fixed condition to create swap partition
The previous condition created the partition if swap is
requested and no volume management is used. But this is
wrong because if any other than the LVM volume management
is used we want the swap space as partition. Only in case
of LVM the swap space is a LVM volume.