Follow up fix for kernel version lookup
When using custom kiwi boot images a shell method named
baseCreateCommonKernelFile was used to create a common
kernel file name in the boot image to allow a simpler
search code in the builder. The search code in the builder
however uses the standard kernel naming schema and with
the latest changes to the kernel version lookup the
common name did not match the matching pattern anymore.
The use of such a common kernel file is obsolete and
the shell method can finally be deleted.