Turn xorriso requirement into a recommend
xorriso was set required on the master python3-kiwi package.
This is not needed since it is only required if an ISO target
will be built. Especially on non x86 architectures the ISO
target is used rarely. Therefore this commit turns xorriso
into a recommended package on non x86 architectures. This is
related to Issue #1325