Changed concept for checking oem resize to be done
The former implementation was based on checking for unallocated
space on the disk compared to the current disk geometry used
in the partition table. This actually is a clean way to check
the condition if the image got already resized to the real
disk geometry. The tools used to check this condition were
sfdisk and sgdisk. The problem is they are different in behavior
and functionality compared across the distributions we support
with kiwi. Thus there was a big risk that the check for
unallocated space failed depending on the target distribution
and release. Therefore I changed the concept to detect the
condition and added an ID string to the root filesystem label.
If the ID string is present we consider it as an initial
deployment, if not we consider the machine deployed and will
not run the resize operation again. On resize kiwi's dracut
code will rewrite the label and remove the ID. Users can
re-activate the resize operation if they stick with the kiwi
oem dracut modules and append the _IMG id string to the
rootfs label.