Make disk allocation check more robust
The tools used to check the disk allocation condition
are sfdisk and sgdisk. The problem is that at least sfdisk
is different in behavior and functionality compared across
the distributions we support with kiwi. In addition
the verification for the msdos table cannot be used to
distinguish between intentionaly wanted free space on
disk and a disk that has not yet been resized. Thus
this commit changes two parts:

a) always report unallocated space available for the msdos
   table to allow to run kiwi's resize code

b) make sure the table type is taken into consideration
   It's important to run the verification based on the
   table type (DOS, GPT) where we know the tools to work. In any
   other case we report the disk to have unallocated space
   and give the resize code a chance