Call grub update wrapper script if present
grub reads its variables from /etc/default/grub but there
are also additional information in /etc/sysconfig. For
moving data to the right place some distributions uses
a wrapper script from perl-BootLoader. This commit checks
if the pbl script exists and checks if it can be called
with --refresh. If yes it's assumed this call moves grub
setup variables into to the right place such that the
follow up grub2-mkconfig call has access to all information
it needs to produce a correct grub.cfg file. This Fixes
bsc#1165960