Refactor dracut disk re-partition library
The kiwi repartition code for msdos and gpt tables was based
on the parted tool. However parted requires to give start and
end addresses to describe a space in the partition table. The
numbers were calculated in kiwi based on cylinders. The mapping
into cylinders is a virtual number that plays no role under
Linux and vanished completely with newer storage technologies.
Tools like sgdisk and sfdisk are better prepared to handle
disk partitioning and also makes this a more straight forward
code in kiwi. Thus this commit refactors the re-partition code
to drop parted and use a dedicated tool depending on the
partition table type. This Fixes #1406