Added custom partition image integration test
There are still people who like many extra static inflexible
entries in a partition table instead of using LVM or filesystem
volume capabilities for sometimes good but often questionable
reasons. In kiwi we intentionally support partition table
entries on a restricted basis but that does not mean you can't
create an image with a highly customized partition table.
However it includes some bits of custom code as part of the
image description and that's what this reference implementation
of an image with custom partitions demonstrates. The image
described here adds three extra partitions, var, var/log and
var/audit. The concept to create custom partitions is based on
specifying a spare partition which is by default created before
the root partition. The spare space can now be sliced into as
many partitions as needed and that needs to be implemented by
the author of the image description. Of course the partition
table itself comes with limitations which has to be respected
depending on the partition table type.