Maintain custom type hinting patch
If we want the upstream code to function on Python 3.4 the
current use of type hints is not possible. The syntax used
in our code is available for Python 3.5 or later. Distributions
like sle12-sp5 still use Python 3.4 and we claim to support
this target. However 3.4 won't be maintained after March
2019 (cf PEP 429) which means upstream we don't care for
this Python version. In order to allow using the latest code
I created a custom patch which we can maintain on the package
level and also port to the enterprise packages if we want
to renew the code base for e.g sle12-sp5 before the distro
reaches EOL.