Remove sed calls to fix /etc/vimrc
An update of vim in Tumbleweed will move /etc/vimrc to /usr/share/vim as part of
the /usr - /etc split. This makes the sed call fail because /etc/vimrc no longer
exists.
However, the fix is not required anymore then, as the vim package dropped the
"syntax on" line from the default vimrc.