Moved sle12 ppc integration test to internal bs
In agreement with IBM the sle12 integration test has been moved
into the internal buildservice. The reason for this change is
a python 3.4 compatibility problem. This version of python is
used in sle12 but would require patching of upstream kiwi in
terms of type hints and annotations to continue to work.
We don't want to cary this patch upstream but in the sle12_kiwi
repository which contains the kiwi used in sle12. Therefore
also the integration test needs to move into the internal
sle12 space.