Fixed root replace for grub menu entries
The root device spec in the grub boot/loader/entries/*.conf
files when building in OBS is not matching the loop mapped
image root device but matches the root device of the OBS worker.
Very strange behavior and of course wrong. Because of that
the expression to replace the wrong device with the correct
one must be a regular expression and cannot be based on the
loop mapped image root device.