Fixed code logic in resize_filesystem method
resize_filesystem runs fs-check on the filesytem prior to
resize. This check however should not be done if the filesystem
got deployed on a ramdisk. For that purpose the _is_ramdisk_device
method exists. The logic in the method as well as the call
were wrong. This in the end lead to a correct logic but is
completely confusing. This commit fixes and simplifies the
_is_ramdisk_device method and corrects the caller logic