Skip filesystem check for XFS prior xfs_grow
running xfs_repair check isn't strictly necessary before resizing,
and in some cases it may even prevent resizing by giving an error
that would be cleared through mounting the fs (e.g. when the fs
wasn't cleanly umounted, and thus letting xfs recover and replay
its journal). Given that xfs can only grow online (while being mounted),
this is sufficient to ensure that the fs is in a state where it
can be resized. This is related to bsc#1174009