Explicitly set start sector for msdos repartition
When using the sfdisk(>=2.26) tool the start and end of partitions
aligns to block-device I/O limits. This means it's hardware specific
what values sfdisk picks. In a repart operation this is however
dangerous for the start sector which must be the same to prevent
corruption of the data in the partition when kiwi resizes it.
Therefore this commit uses the sfdisk dump command to read the
start sector for deleted partitions and uses that start sector
if the same partition gets recreated such that only the end of
the partition changes and gets aligned properly.
This Fixes #1543